Fix: make page.tsx server component
