+++ components/hero/hero.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import ie from "@/public/ichipiroexplorer.png";
 import Image from "next/legacy/image";
 import { useEffect, useState } from "react";
